Changed some UInt32 references to CData::Size. Added missing case statements for references to SValue::kEmpty. Updated all file position references to UInt64. Refactored packet handling from CCodec to SMediaPacket and added additional classes specific to packet handling. Updated CDVIIntealIMAADPCMAudioCodec to use packets instead of raw data. Minor updates to SAudioFormat. Introduced CMediaReader and subclasses and updated CCodec to reference. Reworked Audio Codec interactions to remove stream handling from the codec to CAudioDecoder.

+++ Source/SQLite/CSQLiteStatementPerformer.cpp
@@ -75,7 +75,7 @@ class CSQLiteStatement {
 													// Data
 													sqlite3_bind_blob(statement, i + 1,
 															value.mValue.mData->getBytePtr(),
-															value.mValue.mData->getSize(), SQLITE_STATIC);
+															(int) value.mValue.mData->getSize(), SQLITE_STATIC);
 													break;
 
 												case SSQLiteValue::kLastInsertRowID:
